refactor: split server.ts into api/ + views/ + engine/ modules (#120)
- src/api/ — 11 route files (sessions, feed, teams, config, fleet, asks,
  oracle, federation, worktrees, ui-state, deprecated)
- src/views/ — 5 view files (office, 8bit, war-room, race-track, superman)
- src/engine/ — moved engine.*.ts into engine/ directory
- src/server.ts — now ~80 lines, just mounts routes
- src/paths.ts — shared MAW_ROOT export

3 agents in parallel (api-splitter, views-splitter, engine-reorg).

Closes #120
