+++ src/api/oracle.ts
@@ -1,0 +1,39 @@
+import { Hono } from "hono";
+import { loadConfig } from "../config";
+
+const ORACLE_URL = process.env.ORACLE_URL || loadConfig().oracleUrl;
+
+export const oracleApi = new Hono();
+
+oracleApi.get("/oracle/search", async (c) => {
+  const q = c.req.query("q");
+  if (!q) return c.json({ error: "q required" }, 400);
+  const params = new URLSearchParams({ q, mode: c.req.query("mode") || "hybrid", limit: c.req.query("limit") || "10" });
+  const model = c.req.query("model");
+  if (model) params.set("model", model);
+  try {
+    const res = await fetch(`${ORACLE_URL}/api/search?${params}`);
+    return c.json(await res.json());
+  } catch (e: any) {
+    return c.json({ error: `Oracle unreachable: ${e.message}` }, 502);
+  }
+});
+
+oracleApi.get("/oracle/traces", async (c) => {
+  const limit = c.req.query("limit") || "10";
+  try {
+    const res = await fetch(`${ORACLE_URL}/api/traces?limit=${limit}`);
+    return c.json(await res.json());
+  } catch (e: any) {
+    return c.json({ error: `Oracle unreachable: ${e.message}` }, 502);
+  }
+});
+
+oracleApi.get("/oracle/stats", async (c) => {
+  try {
+    const res = await fetch(`${ORACLE_URL}/api/stats`);
+    return c.json(await res.json());
+  } catch (e: any) {
+    return c.json({ error: `Oracle unreachable: ${e.message}` }, 502);
+  }
+});
