+++ src/views/8bit.ts
@@ -1,0 +1,11 @@
+import { Hono } from "hono";
+import { serveStatic } from "hono/bun";
+import { MAW_ROOT } from "../paths";
+
+export const bitView = new Hono();
+
+bitView.get("/", serveStatic({ root: `${MAW_ROOT}/dist-8bit-office`, path: "/index.html" }));
+bitView.get("/*", serveStatic({
+  root: MAW_ROOT,
+  rewriteRequestPath: (p) => p.replace(/^\/office-8bit/, "/dist-8bit-office"),
+}));
